return type changes based on commit: 45c7e6ca414406487a712ba55afc4e16542775c9

+++ ENOMVG_HFT_2021222.Test/StudentTester.cs
@@ -45,7 +45,7 @@ namespace ENOMVG_HFT_2022231.Test
         [Test]
         public void AvarageAgeTest()
         {
-            int avgA = logic.AvarageAge();
+            double avgA = logic.AvarageAge();
 
             Assert.That(avgA == 12);
         }
